+++ packages/pigeon/platform_tests/test_plugin/android/src/main/kotlin/com/example/test_plugin/TestPlugin.kt
@@ -10,7 +10,6 @@ import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
-import io.flutter.plugin.common.MethodChannel.Result
 
 /**
  * This plugin handles the native side of the integration tests in
@@ -104,87 +103,103 @@ class TestPlugin: FlutterPlugin, HostIntegrationCoreApi {
     return aNullableObject
   }
 
-  override fun noopAsync(callback: () -> Unit) {
-    callback()
+  override fun noopAsync(callback: (Result<Unit>) -> Unit) {
+    callback(Result.success(Unit))
   }
 
-  override fun echoAsyncString(aString: String, callback: (String) -> Unit) {
-    callback(aString)
+  override fun echoAsyncString(aString: String, callback: (Result<String>) -> Unit) {
+    callback(Result.success(aString))
   }
 
-  override fun callFlutterNoop(callback: () -> Unit) {
-    flutterApi!!.noop() { callback() }
-  }
-
-  override fun callFlutterEchoAllTypes(everything: AllTypes, callback: (AllTypes) -> Unit) {
-    flutterApi!!.echoAllTypes(everything) { echo -> callback(echo) }
-  }
-
-  override fun callFlutterSendMultipleNullableTypes(
-    aNullableBool: Boolean?,
-    aNullableInt: Long?,
-    aNullableString: String?,
-    callback: (AllNullableTypes) -> Unit
-  ) {
-    flutterApi!!.sendMultipleNullableTypes(aNullableBool, aNullableInt, aNullableString) {
-      echo -> callback(echo)
+  override fun throwAsyncError(callback: (Result<Any?>) -> Unit) {
+    try {
+      throw Exception("except")
+    } catch (e: Exception) {
+      callback(Result.failure(e))
     }
   }
 
-  override fun callFlutterEchoBool(aBool: Boolean, callback: (Boolean) -> Unit) {
-    flutterApi!!.echoBool(aBool) { echo -> callback(echo) }
+  override fun throwAsyncErrorFromVoid(callback: (Result<Unit>) -> Unit) {
+    try {
+      throw Exception("except")
+    } catch (e: Exception) {
+      callback(Result.failure(e))
+    }
   }
 
-  override fun callFlutterEchoInt(anInt: Long, callback: (Long) -> Unit) {
-    flutterApi!!.echoInt(anInt) { echo -> callback(echo) }
+  override fun callFlutterNoop(callback: (Result<Unit>) -> Unit) {
+    flutterApi!!.noop() { callback(Result.success(Unit)) }
   }
 
-  override fun callFlutterEchoDouble(aDouble: Double, callback: (Double) -> Unit) {
-    flutterApi!!.echoDouble(aDouble) { echo -> callback(echo) }
+  override fun callFlutterEchoAllTypes(everything: AllTypes, callback: (Result<AllTypes>) -> Unit) {
+    flutterApi!!.echoAllTypes(everything) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoString(aString: String, callback: (String) -> Unit) {
-    flutterApi!!.echoString(aString) { echo -> callback(echo) }
+  override fun callFlutterSendMultipleNullableTypes(
+    aNullableBool: Boolean?, 
+    aNullableInt: Long?, 
+    aNullableString: String?, 
+    callback: (Result<AllNullableTypes>) -> Unit
+  ) {
+    flutterApi!!.sendMultipleNullableTypes(aNullableBool, aNullableInt, aNullableString) {
+      echo -> callback(Result.success(echo))
+    }
   }
 
-  override fun callFlutterEchoUint8List(aList: ByteArray, callback: (ByteArray) -> Unit) {
-    flutterApi!!.echoUint8List(aList) { echo -> callback(echo) }
+  override fun callFlutterEchoBool(aBool: Boolean, callback: (Result<Boolean>) -> Unit) {
+    flutterApi!!.echoBool(aBool) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoList(aList: List<Any?>, callback: (List<Any?>) -> Unit) {
-    flutterApi!!.echoList(aList) { echo -> callback(echo) }
+  override fun callFlutterEchoInt(anInt: Long, callback: (Result<Long>) -> Unit) {
+    flutterApi!!.echoInt(anInt) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoMap(aMap: Map<String?, Any?>, callback: (Map<String?, Any?>) -> Unit) {
-    flutterApi!!.echoMap(aMap) { echo -> callback(echo) }
+  override fun callFlutterEchoDouble(aDouble: Double, callback: (Result<Double>) -> Unit) {
+    flutterApi!!.echoDouble(aDouble) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoNullableBool(aBool: Boolean?, callback: (Boolean?) -> Unit) {
-    flutterApi!!.echoNullableBool(aBool) { echo -> callback(echo) }
+  override fun callFlutterEchoString(aString: String, callback: (Result<String>) -> Unit) {
+    flutterApi!!.echoString(aString) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoNullableInt(anInt: Long?, callback: (Long?) -> Unit) {
-    flutterApi!!.echoNullableInt(anInt) { echo -> callback(echo) }
+  override fun callFlutterEchoUint8List(aList: ByteArray, callback: (Result<ByteArray>) -> Unit) {
+    flutterApi!!.echoUint8List(aList) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoNullableDouble(aDouble: Double?, callback: (Double?) -> Unit) {
-    flutterApi!!.echoNullableDouble(aDouble) { echo -> callback(echo) }
+  override fun callFlutterEchoList(aList: List<Any?>, callback: (Result<List<Any?>>) -> Unit){
+    flutterApi!!.echoList(aList) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoNullableString(aString: String?, callback: (String?) -> Unit) {
-    flutterApi!!.echoNullableString(aString) { echo -> callback(echo) }
+  override fun callFlutterEchoMap(aMap: Map<String?, Any?>, callback: (Result<Map<String?, Any?>>) -> Unit) {
+    flutterApi!!.echoMap(aMap) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoNullableUint8List(aList: ByteArray?, callback: (ByteArray?) -> Unit) {
-    flutterApi!!.echoNullableUint8List(aList) { echo -> callback(echo) }
+  override fun callFlutterEchoNullableBool(aBool: Boolean?, callback: (Result<Boolean?>) -> Unit) {
+    flutterApi!!.echoNullableBool(aBool) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoNullableList(aList: List<Any?>?, callback: (List<Any?>?) -> Unit) {
-    flutterApi!!.echoNullableList(aList) { echo -> callback(echo) }
+  override fun callFlutterEchoNullableInt(anInt: Long?, callback: (Result<Long?>) -> Unit) {
+    flutterApi!!.echoNullableInt(anInt) { echo -> callback(Result.success(echo)) }
   }
 
-  override fun callFlutterEchoNullableMap(aMap: Map<String?, Any?>?, callback: (Map<String?, Any?>?) -> Unit) {
-    flutterApi!!.echoNullableMap(aMap) { echo -> callback(echo) }
+  override fun callFlutterEchoNullableDouble(aDouble: Double?, callback: (Result<Double?>) -> Unit) {
+    flutterApi!!.echoNullableDouble(aDouble) { echo -> callback(Result.success(echo)) }
+  }
+
+  override fun callFlutterEchoNullableString(aString: String?, callback: (Result<String?>) -> Unit) {
+    flutterApi!!.echoNullableString(aString) { echo -> callback(Result.success(echo)) }
+  }
+
+  override fun callFlutterEchoNullableUint8List(aList: ByteArray?, callback: (Result<ByteArray?>) -> Unit) {
+    flutterApi!!.echoNullableUint8List(aList) { echo -> callback(Result.success(echo)) }
+  }
+
+  override fun callFlutterEchoNullableList(aList: List<Any?>?, callback: (Result<List<Any?>?>) -> Unit) {
+    flutterApi!!.echoNullableList(aList) { echo -> callback(Result.success(echo)) }
+  }
+
+  override fun callFlutterEchoNullableMap(aMap: Map<String?, Any?>?, callback: (Result<Map<String?, Any?>?>) -> Unit) {
+    flutterApi!!.echoNullableMap(aMap) { echo -> callback(Result.success(echo)) }
   }
 
 }